Update README.md
Adding info on user-controllable changes. We got a comment that people wanted it to be easier to force the calculator to output all csvs individually, so thought the easiest thing would be to point to this in the readme. I seem to remember there were a few other options like this we decided to set to a default but make easy to change.